[UPDATE] #275 Separate block of codes in different files
